% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\docType{data}
\name{color_set}
\alias{color_set}
\title{color_set}
\format{An object of class \code{character} of length 15.}
\usage{
color_set
}
\description{
v0.1.3
add functions of hline and vline
}
\keyword{datasets}
