% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{hline}
\alias{hline}
\title{Add hline to plotly}
\usage{
hline(y = 0, color = "#FF3030", width = 0.5, linetype = "solid")
}
\arguments{
\item{y}{y intercept}

\item{color}{color of vline}

\item{width}{line width, default is 0.5}

\item{linetype}{line type: "solid", "do", "dash", "longdash", "dashdo" or "longdashdot"}
}
\description{
Add hline to plotly
}
