% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{P_preProcess}
\alias{P_preProcess}
\title{Preprocess the input of Plotly_XXX functions and turn them to requirement
of parameters for plotting.}
\usage{
P_preProcess(df, info = NULL, xcolor = NULL, xtext = NULL,
  name_vector = NULL, xaxis_style = NULL, yaxis_style = NULL,
  xGrid = TRUE, yGrid = TRUE, xFactor = FALSE, x_extend = NULL,
  y_extend = NULL, mode = "X")
}
\arguments{
\item{df}{dataframe contain x and y}

\item{info}{vector with the same rows as df to show information for hover}

\item{xcolor}{character vector as the same rows of df for color variable}

\item{xtext}{vector for showing text on plot}

\item{name_vector}{character vector to names x-label, y-label and info}

\item{xGrid}{logical, show x grid}

\item{yGrid}{logical, show y grid}

\item{xFactor}{logical, turn x-variable to factor}

\item{mode}{character, mode of plotly trace, default "X"}

\item{xaxis_stype}{list to layout for setting x-axis style}

\item{yaxis_stype}{list to layout for setting y-axis style}

\item{x_extend(y_extend)}{c(lower_extend_ratio, upper_extend_ratio), coordinate extension}
}
\description{
Preprocess the input of Plotly_XXX functions and turn them to requirement
of parameters for plotting.
}
