% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{P_width}
\alias{P_width}
\title{Count plot width by numbers of x factors}
\usage{
P_width(x, width = c(300, 1000))
}
\arguments{
\item{x}{vecter to count the unique value and than calculate plot width
according the unique value}

\item{width}{a numeric vector with two, min and max pixel of plot width}
}
\description{
Count plot width by numbers of x factors
}
