% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{Plotly_pareto}
\alias{Plotly_pareto}
\title{Function to plot Pareto}
\usage{
Plotly_pareto(category, frequency, title = "Pareto Plot", xlab = "",
  ylab = "發生頻率", topCat = 0)
}
\arguments{
\item{category}{vector of data for category}

\item{frequency}{vector of data as the same length of category}

\item{title}{title of plot}

\item{xlab}{x label}

\item{ylab}{y label}

\item{topCap}{integer, numbers of top categoried to show on plot, 0 for all}
}
\description{
Function to plot Pareto
}
