% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{Plotly_scatter}
\alias{Plotly_scatter}
\title{Function to plotly scatter-line chart}
\usage{
Plotly_scatter(df, info = NULL, xcolor = NULL, xtext = NULL,
  name_vector = NULL, margin = NULL, xaxis_style = NULL,
  yaxis_style = NULL, showLegend = FALSE, xGrid = TRUE,
  yGrid = TRUE, xFactor = FALSE, x_extend = NULL, y_extend = NULL,
  title = "", title_font = NULL, height = NULL,
  textposition = "top", mode = "lines+markers")
}
\arguments{
\item{df}{dataframe contain x and y}

\item{info}{vector with the same rows as df to show information for hover}

\item{xcolor}{character vector as the same rows of df for color variable}

\item{xtext}{vector for showing text on plot}

\item{name_vector}{character vector to names x-label, y-label and info}

\item{margin}{list(l= 1, r= 1, t= 1, b= 1), plot margin}

\item{showLegend}{logical, show legend on plot}

\item{xGrid}{logical, show x grid}

\item{yGrid}{logical, show y grid}

\item{xFactor}{logical, turn x-variable to factor}

\item{title}{title of plot}

\item{title_font}{list of plotly titlefont}

\item{height}{height of plot (width is auto calculated by unique values of x)}

\item{textposition}{posistion of text to show on plot}

\item{mode}{plot_ly mode, "lines+markers", "markers" or "lollipop"}

\item{xaxis_stype}{list to layout for setting x-axis style}

\item{yaxis_stype}{list to layout for setting y-axis style}

\item{x_extend(y_extend)}{c(lower_extend_ratio, upper_extend_ratio), coordinate extension}
}
\description{
Function to plotly scatter-line chart
}
