% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{vline}
\alias{vline}
\title{Add vline to plotly}
\usage{
vline(x = 0, color = "#FF3030", width = 0.5, linetype = "solid")
}
\arguments{
\item{x}{x intercept}

\item{color}{color of vline}

\item{width}{line width, default is 0.5}

\item{linetype}{line type: "solid", "do", "dash", "longdash", "dashdo" or "longdashdot"}
}
\description{
Add vline to plotly
}
