% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{Plotly_pareto}
\alias{Plotly_pareto}
\title{Function to plot Pareto}
\usage{
Plotly_pareto(category, frequency, title = "Pareto Plot", xlab = "",
  ylab = "Frequency", topCat = 0, width = c(300, 1000),
  colors = "#4682B4")
}
\arguments{
\item{category}{vector of data for category}

\item{frequency}{vector of data as the same length of category}

\item{title}{title of plot}

\item{xlab}{x label}

\item{ylab}{y label}

\item{topCat}{integer, numbers of top categoried to show on plot, 0 for all}

\item{width}{a numeric vector with two, min and max pixel of plot width}

\item{colors}{vector of color set for plotting}
}
\description{
Function to plot Pareto
}
